[testsharder] Fix sharding to check for minimum target duration.
fxrev.dev/934227 introduced logic to be able to decrease the target
duration if it caused shards to exceed the max-shard-size. However, some
long tests can exceed the target duration, so we'd want the target
duration to at least be long enough to run that test once. Otherwise we
get an error because we calculate there needs to be 2 shards of the
smaller duration to be able to run that one test. Thus, we should only
update the target duration if it's at least greater than the minimum
required duration. See go/bbid/8765156804665917201 for an example of how
this failed.

Change-Id: Ie8244835bc26e9aa55dc1694ad435be39ab062b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941858
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Danielle Kay <danikay@google.com>