[testing][realm_proxy] Simple prototype with dictionaries
Demonstrates an "echo test" where the "echo protocol" is backed by a
Receiver served by the Test Realm Factory.

This example also includes code to copy the exposed dictionary to the
test, but it doesn't yet do anything since the API that returns the
exposed dictionary is a stub and will always return an empty dict. This
is tracked in fxbug.dev/303719641.

Bug: b/298100106

Change-Id: Iae6274719ba619754429629189eb8b6f74b83ad1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886692
Reviewed-by: Kendal Harland <kjharland@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Gary Bressler <geb@google.com>