[netcfg] Add NormalizedMac dyn NameCompositionRule
Extract the normalized mac calculation into a separate function to be
shared by both the new and existing implementation.

Reframe the name retry policy to allow for the NormalizedMac rule to be
modified and generate a potentially unique name with a different offset.
If all offsets are attempted (the value gets to 255), then we will
surface an error that a unique identifier could not be found.

Test: Ensure existing implementation, notably test_get_usb_255, passes.
Added equivalent test for new implementation and test for properly
unique last bytes.

Bug: 135106
Bug: 135087
Bug: b/309010424
Change-Id: I3959f417022bcd3331b6837b96b8347ddb32e207
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939913
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>