[wlancfg] Temporarily ignore unreachable patterns
The upcoming rust toolchain will be able to detect that an open range
covers everything, so the wildcards will be unncessary. This is a
temporary CL to allow us to gracefully update to the latest rust
toolchain. It will be reverted after the toolchain rolls.

Bug: 309552229
Change-Id: Ifbb551dbed8746422868e7cc04302e73f990a7e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941677
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>