[zircon][sched] Fix scheduler counter trace events.
Fix the scheduler counter trace events to use arguments for the values
and the cpu number for the counter id. Use the top level name "Stats" so
that counters are rendered in Perfetto as "Stats:<type>:<cpu>".

This CL corrects a latent bug that was introduced in the ktrace to fxt
format / API migration.

Change-Id: Id339fbab606077bd7cfc50b497103e0b4e92d642
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/942272
Fuchsia-Auto-Submit: Corey Tabaka <eieio@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>