[rfc] Amend zx_vmo_transfer_data
After much discussion on https://fxrev.dev/856685 and out of band, it
has been determined that it is not possible to guarantee that the
contents of the source and destination VMO are unchanged on failure.
This update to the RFC reflects that consensus, and instead provides a
new guarantee that the call will not fail so long as certain enumerated
conditions hold.

This amendment also removes some constraints to facilitate Starnix's use
of this syscall.

Bug: 126153
Change-Id: I841d96fe69c9115c41c51772daf2e65068a9047d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939973
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>