[gpioimpl][fidl] Add gn arg to enable soft transition
Add a gn arg that OOT gpioimpl drivers can use to enable a soft
transition. New methods will be added to those drivers guarded by
ifdefs based on this arg, which will then be flipped to true in the same
change that adds the methods to the protocol.

Bug: b/306483881
Test: CQ
Change-Id: Ief073e0d3a4d8559a408e2cc5dd3495cae3ffb51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941859
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>