[sdk-tools] Allow images in only one bucket.
Currently, this 'gsutil' function will return an error if any of the possible buckets do not have an image that matches the given version string.

I don't believe that is the expected behavior here. If there are two possible GCS buckets but only one of them has it, I (as the caller of this function) would still like to receive the list of available images.

Bug: b/307571441
Change-Id: Id300263d9c14bd8c98cb86d9010dd10db6a94384
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/942923
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>