[kernel][riscv] sfence.vma when mapping pages
Synchronize hardware when mapping a page in addition to unmap/protect.
It appears that some hardware has some sort of one entry cache of sorts
that causes a page fault loop when mapping a page as a result of a
fault.

According to the architecture this is correct, since the sfence.vma
instruction description is vague enough to cover additional
synchronization that may be needed other than simple TLB flushes.

To avoid a full cross-cpu shootdown on mapping, only shootdown when
updating kernel entries and locally TLB flush when mapping user pages,
which are the vast majority of mappings. Since it's possible another cpu
can spuriously fault on this mapping, add a sfence.vma to the top of the
page fault handler to try to avoid the secondary cpu getting into a page
fault loop.

Change-Id: I29afa3d31ca2415d72211e9f86414b0d246d0da4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940164
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>