[component_manager] Keep resolved component info together
ResolvedInstanceState stores info about a resolved component
that are originally from the Component struct.

Previously it received and stored them as separate fields,
which made the struct and constructor larger, and less readable
in tests that pass a bunch of None values.

Now all the related data is kept together.

Change-Id: I5878ba14afe77aa8d7423d5be118c7ab4aaa26fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941582
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>