[netcfg] Add DeviceClass dyn NameCompositionRule
Extract a device's InterfaceType and map it to the existing prefixes to
produce the string that is used in the construction of an interface
name.

Tested: Added tests to make use of the DeviceClass dynamic rule in
combination with other existing rules.

Bug: 135106
Bug: 135087
Bug: b/309010663
Change-Id: I0f302a36181c7e187806039d8bf05b4795a541a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939355
Fuchsia-Auto-Submit: Antonio Linhart <antoniolinhart@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>