[diagnostics][reader] Enable retrying on empty payloads when fully filtered
When a hierarchy is fully filtered, the Archivist returns an empty
payload and an error explaining that it was fully filtered. Clients
using the ArchiveReader API are observing this error when passing a
selector to request data that has not been yet been published. They
expect ArchiveReader to retry until the data they requested is
present. This is a reasonable expectation in tests, which is the
main use case of ArchiveReader.

This change allows to configure the retry strategy that the
ArchiveReader executes which varies among use cases.

Fixed: 136023
Fixed: b/308979621
Bug: b/309698603
Multiply: session-manager-integration-test
Run-All-Tests: true
Change-Id: I8e517888d5b1915ad6011175608dbec1f950d0a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941152
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>