[netstack] Test root.Routes* in routes.admin tests
Because the existing tests are extremely comprehensive, I only had to
add a single test case to exercise what (at least appears to be) the
entire interface.

For now, the tests are expected to fail for NS2 and NS3, since the
functionality isn't actually implemented yet.

Bug: 130863
Change-Id: Iffa2d3d534139647283421de7633c65d69f4be83
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941913
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>