[04/nn][vim3] Add SMC and HDMITX MMIO to the vim3-display device.
We add all the resources provided by the vim3-hdmi device to
vim3-display pdev device so that the logic that runs in the
current aml-hdmi driver can be migrated to the amlogic-display
driver.

The definitions of the new resources match those in vim3-hdmi.cc.

Test: no-op expected.
Bug: 134691
Change-Id: I6c5c703883c2a8fa183c9c027adf7ea9c227e64c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/944297
Reviewed-by: Victor Costan <costan@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Yilong Li <liyl@google.com>