[kernel][riscv] add support for Zicbom and Zicboz cache instructions
Add kernel support for the Zicbom and Zicboz instructions, needed to
flush the local cache and a fast zero writing option.

Currently enabled all of the time with a cache line size of 64, which is
the standard for RVA22 hardware. QEMU also emulates these instructions,
though most of the time they are a nop.

Bug: 124333

Change-Id: Ic1f133a0fd3d430115fa2546d60c26d21f56d9b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906383
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>