[netstack3] Workaround for UDP-connect to port 0
We've observed real-world instances of applications that rely on
connecting a UDP socket to port 0 to check for presence of IPv4 or
IPv6 internet connections. We need to handle this better with respect
to the socket connected/bound state (as observed via getpeername), but
as a starting point we can check if a route exists to the given
destination and return OK if so.

Bug: 136207
Bug: b/308828371
Test: netstack3_udp_syscall_tests
Change-Id: Id8e4fcdeab78f19d7167faf950998e18b69bd0d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/944652
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>