[seriallistener] Write emulator output from seriallistener.
Since botanist already writes emulator serial output to stdout,
seriallistener discarded the serial output it would read so that we
didn't output the serial logs twice. This CL temporarily disables
writing to stdout from botanist and lets seriallistener write the serial
output to stdout instead. This will make it clearer when the boot test
fails to run because it won't show the serial output unless
seriallistener was able to read it. Also, it should help determine
whether the flakes are due to a race between the two processes trying to
write to stdout or not.

Bug: 135386
Change-Id: I8c19ef0653ab961dad4793fe79b8d82059d6645d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/943941
Commit-Queue: Danielle Kay <danikay@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Danielle Kay <danikay@google.com>