[ld] Add dependency enqueuing to RemoteLoadModule
This CL adds dependency parsing and enqueuing. The RemoteLoadModule does
not have any concept of `preloaded_modules`, so `DecodeDeps` is
simplified to only expect new modules to come across after the first
one (the executable).

The GetVdsoVmo function is moved to its on file in ld::testing so that
it may used by LdRemoteProcessTests.

Bug: 134211
Change-Id: If46197e6cdf263b3f8bf8c291382412a08144d93
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934038
Commit-Queue: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>