[testrunner] Use sshTester's EnsureSinks() if ffx test not enabled.
The FFXTester may be used if ffx is enabled but `ffx test` is only
enabled at experiment level 2. Thus, the way to collect sinks from the
ffx test outputs should also only be enabled when `ffx test` is used.
The coverage builders broke when fxrev.dev/914436 forwarded the ffx
flag to testrunner, causing it to use the FFXTester with experiment
level 1 which means that it was NOT using `ffx test`.

Change-Id: I9d0e1ab6a60c3b8033796d9cab8b2b2142516eab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945979
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Danielle Kay <danikay@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>