[build][ldso] Avoid LTO codegen bug with inline ASM
Clang has a bug whereby module level inline ASM requiring target
features may result in codegen errors that will not cause a link to
fail, but may corrupt the symbol table (e.g., by omitting the function).
Note that this only affects module level ASM, and not all instances of
inline assembly, so the number of places this affects in Fuchsia  is
relatively small. Thus far, it only appears to affect riscv64 in our
codebase, but the bug affects all architectures, and needs to be fixed
in upstream LLVM.

Specifically, we see this causing an issue in dynlink.c. The proposed
workarounds for now are to disable LTO or to move the ASM into assembly
files. This patch adopts the later approach instead of disabling LTO.

More details on the codegen issue can be found on
https://github.com/llvm/llvm-project/issues/67698

Bug: 136220
Test: fx set bringup.riscv64 --variant=lto --with //bundles/boot_tests --auto-dir
Test: fx build
Change-Id: Id96ea732d2b6fd1d46a6f82ec2f91ea857dd9858
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/945114
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Fuchsia-Auto-Submit: Paul Kirth <paulkirth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>