[performance] Automatically generate the most simple benchmark Python drivers
Instead of requiring benchmark authors to write Python host-side code
to drive their benchmarks, we can ask them to call a GN rule that
runs a common parametrized test for them.

In my opinion this improves the DX of writing a benchmark, given that
users can focus on writing their target-side code and then just bring
the benchmark in the build so that it gets automatically executed and
the results published (as long as the expected metrics accept them,
which continue to require Performance+2).

This change implements the minimal functionality, we can expand as we
run into the need of more functionality while we migrate tests to the
new infrastructure.

Fixed: b/308500038
Cq-Include-Trybots: luci.fuchsia.try:terminal.vim3-release
Change-Id: I0e40892a028cd73a1120d46c484a3aea95751b57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938212
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Kevin Cho <chok@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>